AppとRenderを汎用モジュールとして分離
